requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
python-dotenv
google-genai
numpy
